Moving the rest of the client stuff to folder
