+++ server/routes/clients/index.js
@@ -1,0 +1,13 @@
+const express = require('express');
+const router = express.Router();
+
+const { withAsync } = require('../../middleware');
+
+router.get('/', withAsync(require('./getClientsForAgent')));
+router.get('/:id', withAsync(require('./getClient')));
+router.post('/:id/sendEmail', withAsync(require('./sendEmail')));
+router.post('/:id/gift', withAsync(require('./gift')));
+router.put('/:id', withAsync(require('./updateClient')));
+router.post('/', withAsync(require('./createClient')));
+
+module.exports = router;
